Enum conv::errors::GeneralError

pub enum GeneralError {
    Underflow,
    Overflow,
    Unrepresentable,
}

A general error enumeration that subsumes all other conversion errors.

This exists primarily as a "catch-all" for reliably unifying various different kinds of conversion errors.

Variants

Underflow

Input underflowed the target type.

Overflow

Input overflowed the target type.

Unrepresentable

Input was not representable in the target type.