Enum conv::errors::GeneralErrorKind

pub enum GeneralErrorKind {
    NegOverflow,
    PosOverflow,
    Unrepresentable,
}

A general error enumeration that subsumes all other conversion errors, but discards all input payloads the errors may be carrying.

This exists primarily as a "catch-all" for reliably unifying various different kinds of conversion errors, and between different input types.

Variants
NegOverflow

Input was too negative for the target type.

PosOverflow

Input was too positive for the target type.

Unrepresentable

Input was not representable in the target type.