Trait conv::errors::UnwrapOk

pub trait UnwrapOk<T> {
    fn unwrap_ok(self) -> T;
}

Safely unwrap a Result that cannot contain an error.

Required Methods

fn unwrap_ok(self) -> T

Unwraps a Result without possibility of failing.

Technically, this is not necessary; it's provided simply to make user code a little clearer.