Trait conv::errors::UnwrapOrInvalid

pub trait UnwrapOrInvalid {
    type Output;
    fn unwrap_or_invalid(self) -> Self::Output;
}

Unwrap a conversion by replacing a failure with an invalid sentinel value.

Associated Types

type Output

The result of unwrapping.

Required Methods

fn unwrap_or_invalid(self) -> Self::Output

Either unwraps the successfully converted value, or returns the output type's invalid sentinel value.