Trait conv::ApproxWith

pub trait ApproxWith<Dst> {
    fn approx(self) -> Result<Dst, Self::Err> where Self: Sized + ApproxInto<Dst> { ... }
    fn approx_with<Scheme = DefaultApprox>(self) -> Result<Dst, Self::Err> where Self: Sized + ApproxInto<Dst, Scheme>, Scheme: ApproxScheme { ... }
}

This extension trait exists to simplify using approximation implementations.

If there is more than one ApproxFrom implementation for a given type, a simple call to approx_into may not be uniquely resolvable. Due to the position of the scheme parameter (on the trait itself), it is cumbersome to specify which scheme you wanted.

Hence this trait.

Provided Methods

fn approx(self) -> Result<Dst, Self::Err> where Self: Sized + ApproxInto<Dst>

Approximate the subject with the default scheme.

fn approx_with<Scheme = DefaultApprox>(self) -> Result<Dst, Self::Err> where Self: Sized + ApproxInto<Dst, Scheme>, Scheme: ApproxScheme

Approximate the subject with a specific scheme.