Crate conv

This crate provides a number of conversion traits with more specific semantics than those provided by as or From/Into.

The goal with the traits provided here is to be more specific about what generic code can rely on, as well as provide reasonably self-describing alternatives to the standard From/Into traits. For example, the although T: From<U> might be satisfied in generic code, this says nothing about what kind of conversion that represents.

In addition, From/Into provide no facility for a conversion failing, meaning that implementations may need to choose between conversions that may not be valid, or panicking; neither option is appealing in general.

API Stability Notice

The API of this crate is still not entirely decided. In particular, errors may change in the future to carry the value that failed to convert (allowing it to be recovered).

Overview

The following traits are used to define various conversion semantics:

ApproxFrom/ApproxInto - approximate conversions, with selectable approximation scheme (see ApproxScheme).

TryFrom/TryInto - general, potentially failing value conversions.

ValueFrom/ValueInto - exact, value-preserving conversions.

These extension methods are used to make working with potentially failing conversions simpler:

UnwrapOk::unwrap_ok - unwraps results from conversions that cannot fail.

UnwrapOrInf::unwrap_or_inf - saturates to ±∞ on failure.

UnwrapOrInvalid::unwrap_or_invalid - substitutes the target type's "invalid" sentinel value on failure.

UnwrapOrSaturate::unwrap_or_saturate - saturates to the maximum or minimum value of the target type on failure.

A macro is provided to assist in implementing conversions:

TryFrom! - derives an implementation of TryFrom.

If you are implementing your own types, you may also be interested in the traits contained in the misc module.

Reexports

pub use errors::{NoError, GeneralError, Unrepresentable, Underflow, Overflow, FloatError, RangeError, UnwrapOk, UnwrapOrInf, UnwrapOrInvalid, UnwrapOrSaturate};

Modules

errors
This module defines the various error types that can be produced by a failed conversion.

macros
This module provides convenience macros to help with implementing the conversion traits.

misc
This module defines some additional traits not directly tied to conversions.

prelude
Publicly re-exports the most generally useful set of items.

Macros

TryFrom!
See the documentation for the macros module for details.

Enums

DefaultApprox
The "default" approximation scheme. This scheme does whatever would generally be expected of a lossy conversion, assuming no additional context or instruction is given.

Wrapping
This scheme is used to convert a value by "wrapping" it into a narrower range.

Traits

ApproxFrom
This trait is used to perform a conversion that is permitted to approximate the result, but not to wrap or saturate the result to fit into the destination type's representable range.

ApproxInto
This is the dual of ApproxFrom; see that trait for information.

ApproxScheme
This trait is used to mark approximation scheme types.

ApproxWith
This extension trait exists to simplify using approximation implementations.

TryFrom
This trait is used to perform a conversion between different semantic types which might fail.

TryInto
This is the dual of TryFrom; see that trait for information.

ValueFrom
This trait is used to perform an exact, value-preserving conversion.

ValueInto
This is the dual of ValueFrom; see that trait for information.