Enum conv::DefaultApprox

pub enum DefaultApprox {}

The “default” approximation scheme. This scheme does whatever would generally be expected of a lossy conversion, assuming no additional context or instruction is given.

This is a double-edged sword: it has the loosest semantics, but is far more likely to exist than more complicated approximation schemes.