Enum conv::Wrapping

pub enum Wrapping {}

This scheme is used to convert a value by "wrapping" it into a narrower range.

In abstract, this can be viewed as the opposite of rounding: rather than preserving the most significant bits of a value, it preserves the least significant bits of a value.