Module conv::errors

This module defines the various error types that can be produced by a failed conversion.

In addition, it also defines some extension traits to make working with failable conversions more ergonomic (see the Unwrap* traits).

Structs

NegOverflow
Indicates that the conversion failed due to a negative overflow.

PosOverflow
Indicates that the conversion failed due to a positive overflow.

Unrepresentable
Indicates that the conversion failed because the value was not representable.

Enums

FloatError
Indicates that a conversion from a floating point type failed.

GeneralError
A general error enumeration that subsumes all other conversion errors.

GeneralErrorKind
A general error enumeration that subsumes all other conversion errors, but discards all input payloads the errors may be carrying.

NoError
Indicates that it is not possible for the conversion to fail.

RangeError
Indicates that a conversion failed due to a range error.

RangeErrorKind
Indicates that a conversion failed due to a range error.

Traits

Saturate
Saturates a Result.

UnwrapOk
Safely unwrap a Result that cannot contain an error.

UnwrapOrInf
Unwrap a conversion by saturating to infinity.

UnwrapOrInvalid
Unwrap a conversion by replacing a failure with an invalid sentinel value.

UnwrapOrSaturate
Unwrap a conversion by saturating.