Trait conv::TryFrom

pub trait TryFrom<Src>: Sized {
    type Err: Error;

    // Required method
    fn try_from(src: Src) -> Result<Self, Self::Err>;
}

This trait is used to perform a conversion between different semantic types which might fail.

Where possible, prefer implementing this trait over TryInto, but prefer using TryInto for generic constraints.

Details

Typically, this should be used in cases where you are converting between values whose ranges and/or representations only partially overlap. That the conversion may fail should be a reasonably expected outcome. A standard example of this is converting from integers to enums of unitary variants.

Required Associated Types

type Err: Error

The error type produced by a failed conversion.

Required Methods

fn try_from(src: Src) -> Result<Self, Self::Err>

Convert the given value into the subject type.